requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
slack-sdk==3.20.2
python-dotenv
